requirements.txt
appdirs==1.4.4
dpkt==1.9.8
lxml==5.3.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
pygeoip==0.3.2
pyshark==0.6
python-dateutil==2.9.0.post0
pytz==2025.1
scapy==2.6.1
six==1.17.0
termcolor==2.5.0
tzdata==2025.1
yara==1.7.7
yara-python==4.5.1
